backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.12.0
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
cryptography==41.0.7
PyJWT==2.11.0
djangorestframework-simplejwt==5.5.1
django-filter==23.5
drf-spectacular==0.26.3
mysqlclient==2.2.0